+++ matrix_io/malos/cli.py
@@ -94,12 +94,16 @@ async def status_handler(malos_driver):
     """
 
     type_mapping = {
-        driver_pb2.Status.NOT_DEFINED: "Not Defined",
+        driver_pb2.Status.MESSAGE_TYPE_NOT_DEFINED: "Not Defined",
         driver_pb2.Status.STARTED: "Started",
+        driver_pb2.Status.STOPPED: "Stopped",
         driver_pb2.Status.CONFIG_RECEIVED: "Config Received",
         driver_pb2.Status.COMMAND_EXECUTED: "Command Executed",
-        driver_pb2.Status.ERROR: "Error",
-        driver_pb2.Status.WARNING: "Warning"
+        driver_pb2.Status.STATUS_CRITICAL: "Critical log",
+        driver_pb2.Status.STATUS_ERROR: "Error log",
+        driver_pb2.Status.STATUS_WARNING: "Warning log",
+        driver_pb2.Status.STATUS_INFO: "Info log",
+        driver_pb2.Status.STATUS_DEBUG: "Debug log"
     }
 
     async for msg in malos_driver.get_status():